re-implement seq_prog.c by re-using the program list in seqCommands.c
This avoids a race condition when creating the global progList (in
seq_prog.c) and also avoids code duplication. The global variables in
seq_prog.c have been removed. Instead we extend the sequencer program
list in seqCommands.c so that for each program it also stores a list
of program instances, each one protected by a mutex. Access to this
list is restricted to a single generic traversal procedure which
automatically applies the lock around each call to the supplied
traversee. Since the program list in seqCommands.c is created before
multithreading starts there is no race condition.

Also added an int return value to the traversees, to indicate whether
traversal is supposed to stop (TRUE) or continue (FALSE).