added prog and ss init, removed ss entry and exit functions
Initialization of variables is now done in the added init
functions, separate from the program entry function. This
is necessary because variable initialization must be done
before channels are connected to avoid races, but now we
connect before the program entry function is called, so
we can use pv functions in the entry handler.
Also refactored generation of global prog entry/exit and
simplified other function generators; renamed generated
prog and ss functions.