snc: fixed (and simplified) type parsing
This was completely broken, since the constructors were
applied in the wrong order. Fixed by building the type
expression in lifo order, so we need additional parent
pointer in struct type. Finding the base type is now a
constant time operation because we abuse parent pointer
for this.
Also moved type code generation to var_types.
Also renamed Var.value to Var.init.