seq: use locks for shared buffer
The previous implementation largely avoids taking locks
but at the cost of being sensitive to thread priorities.
I decided this is too brittle; logical correctness should
not depend on thread priorities or fair scheduling.
A second change closely interwoven with this is to mark
a variable dirty even if its shared buffer is written by a
pvPut (an anonymous PV in safe mode). This removes a special
case and makes anonymous PVs behave more like named ones.