added exactly sized integer types to compiler and runtime
These types are named like the standard C99 types, but are
implemented using epicsTypes.h since the C99 inttypes.h cannot
be assumed to be available on all platforms. We only add
integer types with sizes up to 32, since 64 bit types are not
supported by CA and epicsTypes defines them only if C99 is enabled.