added regression test for the new fixed size integer types
Also test that pvGet/pvPut works with PV of unsigned native
type and requests that are of smaller, equal, and larger size
and signed as well as unsigned work (or fail) as expected.
Note that for unsigned integer variables, the sequencer
has always used signed request types of equal bit size.