cast arguments to isalpha etc to unsigned char
Some compilers implement the functions from ctype.h as macros
that use the argument to index into an array and give a
warning if char is used as an array index. Although the man page
lists the declaration of these functions as taking an int,
it also says the argument should in fact be an unsigned char.
So we cast to be on the safe side.