seq: fixed a bug in task main loop that caused task deadlocks
This bug was introduced when refactoring the seq_getTimeout procedure.
With the changed implementation of seq_getTimeout, it is no longer
correct to check that the returned delay is > 0. In contrast to the
old implementation, failing the check means we run into epicsEventWait
(no timeout) and then wait forever. Returning a delay of 0.0 happens
only if (when calculating the event timeout) the time elapsed is
already greater than the time we would have to wait, and thus happens
only on slow machines.