snc: fixed compiler crash in case of excess PVs given in an assign clause
Funny, I had an assertion in the code that should have caught the bug
but unfortunately the assertion itself had an off-by-one error and so
didn't fire. So I had to debug this the hard way. The compiler will now
give a warning message saying that it discards the excess PV names.