seq: fix pvPut/pvGet hanger on disconnect
The connection handler now releases all semaphores waiting for an
event on this PV. Furthermore, the pvGet/pvPut[Complete] built-in
functions check the connected status whenever they successfully waited
for such a semaphore.

This patch also fixes a few problems with error handling in these
procedures; for instance, the SYNC versions released semaphores even
in case of timeout, which is definitely wrong.