snc: fixed rules for generating parser and lexer
With parallel make (-j) the snl.c and snl.h targets were built
twice which could cause building to fail. With the changed rules
I ran into a make bug and had to add an empty recipy where it
should not be needed. Also needed to work around a bug in the base
rules (no explicit rule exists that generates the .d files).