snc: new syntax nodes for struct or union member selection
Members are no longer identified with variables in the syntax. Member
selection "operators" ('.' and '->') are no longer treated as binary
operators. Instead, member selection gets a new syntax node type E_SELECT,
and members (only allowed as rhs of a member selection) are not E_VAR, but
E_MEMBER. This all works because members are not first class in C, that is
"x.(name)" is a syntax error.

This is a simple solution for the recurring annoyance of snc issuing
"undefined variable" warnings.

Declaring members in a foreign declaration becomes obsolete.