moved non-API internals from seqCom.h to new seq_snc.h
Another one of the long overdue re-factorings. The new header file
seq_snc.h contains all the definitions that are shared between the
snc generated code and the run-time library. What remains in seqCom.h
is the public API for C code that wants to interact with an SNL program.