docs/snc: check syntax productions in the reference for consistency
We no longer include the syntax summary generated from the parser input
file in the reference. This was unreadable anyway. Instead, check that
the syntax productions in the reference are consistent with the
generated ones and complain during build if that is not the case.

Along the way, renamed the generated file from bnf.inc to Syntax.txt and
install it into the html install location; also renamed about.inc to
About.txt, adding both to unused_docs in the sphinx configuration.