snc: removed header file generation
The new SNL functions make a lot of embedded C code, especially C
function definitions, obsolete. This means the header file is not as
useful as I thought. On the other hand, function definitions may depend
on stuff that has been declared before. So it becomes more important to
keep the order of definitions in the output as similar as possible to
that in the input file (the variables struct being the most notable
exception). Grouping some stuff in a header file and other stuff in the
C file makes this almost impossible.