snc: completed implementation of the new multi-PV behaviour
The new boolean member multi_pv of struct func_symbol represents whether
multi-pv arguments are allowed by default, i.e with -p in effect. The code
generator takes this into account when deciding what to do for a multi-PV
argument. The behaviour is now exactly as described in the release notes.