dont add a prefix to names of event flags, generate static const, not #define
If the name of an event flag clashes with an already defined C identifier,
we will have a problem anyway. Besides, variable names aren't mangeled with
a prefix either. Also, in the next version (with pv types) event flags will
most probably be a part of the global variable struct and thus also not
mangled. Lastly, it seems a bad idea to make escaped C code depend on the
naming convention.