removed p option, add pvArrayVariants for some builtins
This patch also re-structures the mechanism with which built-in functions
are handled in the compiler. It also reverts most of the changes in the C
API to increase compatibility. For instance, pVar and ssId are now
"constants", not functions. For seq_pvGet and seq_pvPut there are now ...Tmo
variants with the extra timeout parameter.