fix: avoid deadlock when exiting from a program
This is another instance of the same we had before with pvAssign, only this
time it is seq_disconnect which calls pvVarDestroy which in turn calls
ca_clear_channel. The deadlock is with a pending CA callback which also
tries to take the program lock. The solution is the same: unlock before
calling pvVarDestroy, and re-take the lock afterwards.