cleaned up the prim_types mess, snc no longer depends on seq library
Making the compiler depend on the seq library was a bad idea, since the
names of the primitive types were the only thing that snc needed from seq.

Instead I added a new src/common subdir that installs (and for seq_release.h
generates) the header files that are shared between the compiler and the
runtime. The epicsShare stuff has been removed from seq_prim_types.h. A new
preprocessor symbol (declare_prim_type_names) makes sure that the
prim_type_name and prim_type_tag_name variables are visible only inside the
runtime library and the compiler and not  in generated sequencer programs.