seq: fixed a possible race condition with pvAssign
The problem concerns access to the dbch member of CHAN: the runtime pvAssign
may free this pointer (and also sets it to NULL); on the other hand, the CA
callbacks need to access it, or else must return immediately. There was a
test that checks if dbch is NULL but it was not locked against concurrent
access from pvAssign, which could lead to an assertion failure or worse to
invalid memory accesses.