Fix build problems in examples/ on Windows
The registerRecordDeviceDriver.pl script generates code that on Windows
requires symbols named in the DBD file to be linked from an DLL, not from a
directly-linked source file. When main() is directly linked with an object
file containing a registrar entry, we can call that registrar code directly
from main() as done here instead of using the generated code.

These build errors may only appear when using a MinGW compiler, either
native or as a cross-build running on another OS.